wgsim: コンピュートシェーダーの結果をPNGに書き出すためのヘルパーを提供
