+++ lib/wgsim/Cargo.toml
@@ -9,6 +9,7 @@ cgmath            = "0.18.0"
 futures-intrusive = "0.5.0"
 gif               = "0.13.1"
 indicatif         = "0.17.8"
+png               = "0.17.14"
 pollster          = "0.3.0"
 wgpu              = "22.1.0"
 winit             = "0.30.5"
