+++ tutorial/compute_visualize_workgroup_local/Cargo.toml
@@ -4,9 +4,6 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-env_logger        = "0.11.5"
-wgpu              = "22.1.0"
-wgsim             = { path = "../../lib/wgsim" }
-futures-intrusive = "0.5.0"
-bytemuck          = "1.19.0"
-png               = "0.17.14"
+env_logger = "0.11.5"
+wgpu       = "22.1.0"
+wgsim      = { path = "../../lib/wgsim" }
